Fixed based on feedback of PR and removed deploy, list files. Resolved comments
