+++ src/commands/index.js
@@ -109,12 +109,8 @@ module.exports = function(client) {
   client.target.clear = loadCommand("target-clear");
   client.target.remove = loadCommand("target-remove");
   client.use = loadCommand("use");
-
   client.remoteconfig = {};
   client.remoteconfig.get = loadCommand("remoteconfig-get");
-  client.remoteconfig.versions = {};
-  client.remoteconfig.versions.list = loadCommand("remoteconfig-versions-list");
-  // client.remoteconfig.deploy = loadCommand("remoteconfig-deploy");
 
   return client;
 };
